sweet alert y sharedPreferences
