--- app/src/main/java/Models/DatosHorario.java
@@ -1,259 +1,0 @@
-package Models;
-
-import java.util.Date;
-
-public class DatosHorario {
-
-    String horC_DIA;
-    String horC_SIGLA_DIA;
-    String horC_USU_REG;
-    Date horD_FEC_REG;
-    String horC_USU_MOD;
-    Date horD_FEC_MOD;
-    Boolean horC_ESTADO;
-    int empN_ID_EMPRESA;
-    String empN_ID_EMPRESA_TXT;
-    String mensaje;
-    int areN_ID_PADRE;
-    int areN_ID_HIJO;
-    String areN_ID_PADRE_TXT;
-    String areN_ID_HIJO_TXT;
-    int tiphorN_ID_TIPO_HORARIO;
-    String tiphorN_ID_TIPO_HORARIO_TXT;
-    int hodN_ID_HORARIO_DETALLE;
-    int horN_ID_HORARIO;
-    String hodC_NOMBRE;
-    String hodC_SIGLA;
-    String hodC_HORA;
-    String hodC_INICIO;
-    String hodC_FIN;
-    String hodC_USU_REG;
-    Date hodD_FEC_REG;
-    String hodC_USU_MOD;
-    Date hodD_FEC_MOD;
-    Boolean hodC_ESTADO;
-
-    public String getHorC_DIA() {
-        return horC_DIA;
-    }
-
-    public void setHorC_DIA(String horC_DIA) {
-        this.horC_DIA = horC_DIA;
-    }
-
-    public String getHorC_SIGLA_DIA() {
-        return horC_SIGLA_DIA;
-    }
-
-    public void setHorC_SIGLA_DIA(String horC_SIGLA_DIA) {
-        this.horC_SIGLA_DIA = horC_SIGLA_DIA;
-    }
-
-    public String getHorC_USU_REG() {
-        return horC_USU_REG;
-    }
-
-    public void setHorC_USU_REG(String horC_USU_REG) {
-        this.horC_USU_REG = horC_USU_REG;
-    }
-
-    public Date getHorD_FEC_REG() {
-        return horD_FEC_REG;
-    }
-
-    public void setHorD_FEC_REG(Date horD_FEC_REG) {
-        this.horD_FEC_REG = horD_FEC_REG;
-    }
-
-    public String getHorC_USU_MOD() {
-        return horC_USU_MOD;
-    }
-
-    public void setHorC_USU_MOD(String horC_USU_MOD) {
-        this.horC_USU_MOD = horC_USU_MOD;
-    }
-
-    public Date getHorD_FEC_MOD() {
-        return horD_FEC_MOD;
-    }
-
-    public void setHorD_FEC_MOD(Date horD_FEC_MOD) {
-        this.horD_FEC_MOD = horD_FEC_MOD;
-    }
-
-    public Boolean getHorC_ESTADO() {
-        return horC_ESTADO;
-    }
-
-    public void setHorC_ESTADO(Boolean horC_ESTADO) {
-        this.horC_ESTADO = horC_ESTADO;
-    }
-
-    public int getEmpN_ID_EMPRESA() {
-        return empN_ID_EMPRESA;
-    }
-
-    public void setEmpN_ID_EMPRESA(int empN_ID_EMPRESA) {
-        this.empN_ID_EMPRESA = empN_ID_EMPRESA;
-    }
-
-    public String getEmpN_ID_EMPRESA_TXT() {
-        return empN_ID_EMPRESA_TXT;
-    }
-
-    public void setEmpN_ID_EMPRESA_TXT(String empN_ID_EMPRESA_TXT) {
-        this.empN_ID_EMPRESA_TXT = empN_ID_EMPRESA_TXT;
-    }
-
-    public String getMensaje() {
-        return mensaje;
-    }
-
-    public void setMensaje(String mensaje) {
-        this.mensaje = mensaje;
-    }
-
-    public int getAreN_ID_PADRE() {
-        return areN_ID_PADRE;
-    }
-
-    public void setAreN_ID_PADRE(int areN_ID_PADRE) {
-        this.areN_ID_PADRE = areN_ID_PADRE;
-    }
-
-    public int getAreN_ID_HIJO() {
-        return areN_ID_HIJO;
-    }
-
-    public void setAreN_ID_HIJO(int areN_ID_HIJO) {
-        this.areN_ID_HIJO = areN_ID_HIJO;
-    }
-
-    public String getAreN_ID_PADRE_TXT() {
-        return areN_ID_PADRE_TXT;
-    }
-
-    public void setAreN_ID_PADRE_TXT(String areN_ID_PADRE_TXT) {
-        this.areN_ID_PADRE_TXT = areN_ID_PADRE_TXT;
-    }
-
-    public String getAreN_ID_HIJO_TXT() {
-        return areN_ID_HIJO_TXT;
-    }
-
-    public void setAreN_ID_HIJO_TXT(String areN_ID_HIJO_TXT) {
-        this.areN_ID_HIJO_TXT = areN_ID_HIJO_TXT;
-    }
-
-    public int getTiphorN_ID_TIPO_HORARIO() {
-        return tiphorN_ID_TIPO_HORARIO;
-    }
-
-    public void setTiphorN_ID_TIPO_HORARIO(int tiphorN_ID_TIPO_HORARIO) {
-        this.tiphorN_ID_TIPO_HORARIO = tiphorN_ID_TIPO_HORARIO;
-    }
-
-    public String getTiphorN_ID_TIPO_HORARIO_TXT() {
-        return tiphorN_ID_TIPO_HORARIO_TXT;
-    }
-
-    public void setTiphorN_ID_TIPO_HORARIO_TXT(String tiphorN_ID_TIPO_HORARIO_TXT) {
-        this.tiphorN_ID_TIPO_HORARIO_TXT = tiphorN_ID_TIPO_HORARIO_TXT;
-    }
-
-    public int getHodN_ID_HORARIO_DETALLE() {
-        return hodN_ID_HORARIO_DETALLE;
-    }
-
-    public void setHodN_ID_HORARIO_DETALLE(int hodN_ID_HORARIO_DETALLE) {
-        this.hodN_ID_HORARIO_DETALLE = hodN_ID_HORARIO_DETALLE;
-    }
-
-    public int getHorN_ID_HORARIO() {
-        return horN_ID_HORARIO;
-    }
-
-    public void setHorN_ID_HORARIO(int horN_ID_HORARIO) {
-        this.horN_ID_HORARIO = horN_ID_HORARIO;
-    }
-
-    public String getHodC_NOMBRE() {
-        return hodC_NOMBRE;
-    }
-
-    public void setHodC_NOMBRE(String hodC_NOMBRE) {
-        this.hodC_NOMBRE = hodC_NOMBRE;
-    }
-
-    public String getHodC_SIGLA() {
-        return hodC_SIGLA;
-    }
-
-    public void setHodC_SIGLA(String hodC_SIGLA) {
-        this.hodC_SIGLA = hodC_SIGLA;
-    }
-
-    public String getHodC_HORA() {
-        return hodC_HORA;
-    }
-
-    public void setHodC_HORA(String hodC_HORA) {
-        this.hodC_HORA = hodC_HORA;
-    }
-
-    public String getHodC_INICIO() {
-        return hodC_INICIO;
-    }
-
-    public void setHodC_INICIO(String hodC_INICIO) {
-        this.hodC_INICIO = hodC_INICIO;
-    }
-
-    public String getHodC_FIN() {
-        return hodC_FIN;
-    }
-
-    public void setHodC_FIN(String hodC_FIN) {
-        this.hodC_FIN = hodC_FIN;
-    }
-
-    public String getHodC_USU_REG() {
-        return hodC_USU_REG;
-    }
-
-    public void setHodC_USU_REG(String hodC_USU_REG) {
-        this.hodC_USU_REG = hodC_USU_REG;
-    }
-
-    public Date getHodD_FEC_REG() {
-        return hodD_FEC_REG;
-    }
-
-    public void setHodD_FEC_REG(Date hodD_FEC_REG) {
-        this.hodD_FEC_REG = hodD_FEC_REG;
-    }
-
-    public String getHodC_USU_MOD() {
-        return hodC_USU_MOD;
-    }
-
-    public void setHodC_USU_MOD(String hodC_USU_MOD) {
-        this.hodC_USU_MOD = hodC_USU_MOD;
-    }
-
-    public Date getHodD_FEC_MOD() {
-        return hodD_FEC_MOD;
-    }
-
-    public void setHodD_FEC_MOD(Date hodD_FEC_MOD) {
-        this.hodD_FEC_MOD = hodD_FEC_MOD;
-    }
-
-    public Boolean getHodC_ESTADO() {
-        return hodC_ESTADO;
-    }
-
-    public void setHodC_ESTADO(Boolean hodC_ESTADO) {
-        this.hodC_ESTADO = hodC_ESTADO;
-    }
-}
