+++ app/src/main/java/Models/HorarioViewModel.java
@@ -1,16 +1,188 @@
 package Models;
 
+import java.util.Date;
 import java.util.List;
 
 public class HorarioViewModel {
+    String horC_DIA;
+    String horC_SIGLA_DIA;
+    Boolean horC_ESTADO;
+    int empN_ID_EMPRESA;
+    String empN_ID_EMPRESA_TXT;
+    String mensaje;
+    int areN_ID_PADRE;
+    int areN_ID_HIJO;
+    String areN_ID_PADRE_TXT;
+    String areN_ID_HIJO_TXT;
+    int tiphorN_ID_TIPO_HORARIO;
+    String tiphorN_ID_TIPO_HORARIO_TXT;
+    int hodN_ID_HORARIO_DETALLE;
+    int horN_ID_HORARIO;
+    String hodC_NOMBRE;
+    String hodC_SIGLA;
+    String hodC_HORA;
+    String hodC_INICIO;
+    String hodC_FIN;
+    Boolean hodC_ESTADO;
 
-    DatosHorario datosHorario;
-
-    public DatosHorario getDatosHorario() {
-        return datosHorario;
+    public String getHorC_DIA() {
+        return horC_DIA;
     }
 
-    public void setDatosHorario(DatosHorario datosHorario) {
-        this.datosHorario = datosHorario;
+    public void setHorC_DIA(String horC_DIA) {
+        this.horC_DIA = horC_DIA;
+    }
+
+    public String getHorC_SIGLA_DIA() {
+        return horC_SIGLA_DIA;
+    }
+
+    public void setHorC_SIGLA_DIA(String horC_SIGLA_DIA) {
+        this.horC_SIGLA_DIA = horC_SIGLA_DIA;
+    }
+
+    public Boolean getHorC_ESTADO() {
+        return horC_ESTADO;
+    }
+
+    public void setHorC_ESTADO(Boolean horC_ESTADO) {
+        this.horC_ESTADO = horC_ESTADO;
+    }
+
+    public int getEmpN_ID_EMPRESA() {
+        return empN_ID_EMPRESA;
+    }
+
+    public void setEmpN_ID_EMPRESA(int empN_ID_EMPRESA) {
+        this.empN_ID_EMPRESA = empN_ID_EMPRESA;
+    }
+
+    public String getEmpN_ID_EMPRESA_TXT() {
+        return empN_ID_EMPRESA_TXT;
+    }
+
+    public void setEmpN_ID_EMPRESA_TXT(String empN_ID_EMPRESA_TXT) {
+        this.empN_ID_EMPRESA_TXT = empN_ID_EMPRESA_TXT;
+    }
+
+    public String getMensaje() {
+        return mensaje;
+    }
+
+    public void setMensaje(String mensaje) {
+        this.mensaje = mensaje;
+    }
+
+    public int getAreN_ID_PADRE() {
+        return areN_ID_PADRE;
+    }
+
+    public void setAreN_ID_PADRE(int areN_ID_PADRE) {
+        this.areN_ID_PADRE = areN_ID_PADRE;
+    }
+
+    public int getAreN_ID_HIJO() {
+        return areN_ID_HIJO;
+    }
+
+    public void setAreN_ID_HIJO(int areN_ID_HIJO) {
+        this.areN_ID_HIJO = areN_ID_HIJO;
+    }
+
+    public String getAreN_ID_PADRE_TXT() {
+        return areN_ID_PADRE_TXT;
+    }
+
+    public void setAreN_ID_PADRE_TXT(String areN_ID_PADRE_TXT) {
+        this.areN_ID_PADRE_TXT = areN_ID_PADRE_TXT;
+    }
+
+    public String getAreN_ID_HIJO_TXT() {
+        return areN_ID_HIJO_TXT;
+    }
+
+    public void setAreN_ID_HIJO_TXT(String areN_ID_HIJO_TXT) {
+        this.areN_ID_HIJO_TXT = areN_ID_HIJO_TXT;
+    }
+
+    public int getTiphorN_ID_TIPO_HORARIO() {
+        return tiphorN_ID_TIPO_HORARIO;
+    }
+
+    public void setTiphorN_ID_TIPO_HORARIO(int tiphorN_ID_TIPO_HORARIO) {
+        this.tiphorN_ID_TIPO_HORARIO = tiphorN_ID_TIPO_HORARIO;
+    }
+
+    public String getTiphorN_ID_TIPO_HORARIO_TXT() {
+        return tiphorN_ID_TIPO_HORARIO_TXT;
+    }
+
+    public void setTiphorN_ID_TIPO_HORARIO_TXT(String tiphorN_ID_TIPO_HORARIO_TXT) {
+        this.tiphorN_ID_TIPO_HORARIO_TXT = tiphorN_ID_TIPO_HORARIO_TXT;
+    }
+
+    public int getHodN_ID_HORARIO_DETALLE() {
+        return hodN_ID_HORARIO_DETALLE;
+    }
+
+    public void setHodN_ID_HORARIO_DETALLE(int hodN_ID_HORARIO_DETALLE) {
+        this.hodN_ID_HORARIO_DETALLE = hodN_ID_HORARIO_DETALLE;
+    }
+
+    public int getHorN_ID_HORARIO() {
+        return horN_ID_HORARIO;
+    }
+
+    public void setHorN_ID_HORARIO(int horN_ID_HORARIO) {
+        this.horN_ID_HORARIO = horN_ID_HORARIO;
+    }
+
+    public String getHodC_NOMBRE() {
+        return hodC_NOMBRE;
+    }
+
+    public void setHodC_NOMBRE(String hodC_NOMBRE) {
+        this.hodC_NOMBRE = hodC_NOMBRE;
+    }
+
+    public String getHodC_SIGLA() {
+        return hodC_SIGLA;
+    }
+
+    public void setHodC_SIGLA(String hodC_SIGLA) {
+        this.hodC_SIGLA = hodC_SIGLA;
+    }
+
+    public String getHodC_HORA() {
+        return hodC_HORA;
+    }
+
+    public void setHodC_HORA(String hodC_HORA) {
+        this.hodC_HORA = hodC_HORA;
+    }
+
+    public String getHodC_INICIO() {
+        return hodC_INICIO;
+    }
+
+    public void setHodC_INICIO(String hodC_INICIO) {
+        this.hodC_INICIO = hodC_INICIO;
+    }
+
+    public String getHodC_FIN() {
+        return hodC_FIN;
+    }
+
+    public void setHodC_FIN(String hodC_FIN) {
+        this.hodC_FIN = hodC_FIN;
+    }
+
+    public Boolean getHodC_ESTADO() {
+        return hodC_ESTADO;
+    }
+
+    public void setHodC_ESTADO(Boolean hodC_ESTADO) {
+        this.hodC_ESTADO = hodC_ESTADO;
     }
 }
+
