+++ app/src/main/java/com/example/asistencia/ui/home/HomeFragment.java
@@ -18,25 +18,36 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
 
+import com.android.volley.AuthFailureError;
 import com.android.volley.Request;
 import com.android.volley.RequestQueue;
 import com.android.volley.Response;
 import com.android.volley.VolleyError;
+import com.android.volley.toolbox.JsonArrayRequest;
 import com.android.volley.toolbox.JsonObjectRequest;
+import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
+import com.example.asistencia.LoginActivity;
 import com.example.asistencia.R;
 import com.google.gson.Gson;
 import com.google.zxing.integration.android.IntentIntegrator;
 import com.google.zxing.integration.android.IntentResult;
 
+import org.json.JSONArray;
 import org.json.JSONObject;
 
+import java.net.URL;
+import java.text.DateFormat;
+import java.text.ParsePosition;
+import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
 import Models.HorarioViewModel;
+import Models.LoginViewModel;
+import cn.pedant.SweetAlert.SweetAlertDialog;
 
 public class HomeFragment extends Fragment {
 
@@ -44,13 +55,12 @@ public class HomeFragment extends Fragment {
     Button btnCodigo;
     private static RequestQueue queue;
     private static final String ApiURL = "http://b3rs3rk3r-002-site2.htempurl.com/api/";
-    HorarioViewModel horariosArray = new HorarioViewModel();
+    HorarioViewModel[] horariosArray;
 
     String HoraInicio = "";
     String HoraFin = "";
     String Sigla = "";
     String Tipo = "";
-
 
     public HomeFragment(){
     }
@@ -61,7 +71,7 @@ public class HomeFragment extends Fragment {
         btnCodigo = view.findViewById(R.id.btncodigo);
 
         //QUITAR
-        Toast.makeText(getContext(), getFromSharedPreferences("idusuario"), Toast.LENGTH_SHORT).show();
+        //Toast.makeText(getContext(), getFromSharedPreferences("idusuario"), Toast.LENGTH_SHORT).show();
 
         btnCodigo.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -106,46 +116,65 @@ public class HomeFragment extends Fragment {
         }
     }
 
-
     private boolean marcarAsistencia(int _id) {
-        Boolean status = true;
         if (!ValidarHorario(_id)) {
+            new SweetAlertDialog(getActivity(), SweetAlertDialog.ERROR_TYPE).setTitleText("Fuera de horario permitido").show();
+            return false;
+        } else {
 
-            status = false;
+            return true;
         }
-        return status;
     }
 
     private boolean ValidarHorario(int id) {
-        Boolean status = true;
-        HoraInicio = ""; HoraFin = ""; Sigla = ""; Tipo = "";
-        //RECUPERANDO HORARIO
-        getHorarios(id);
-        return status;
+        try {
+            //RECUPERANDO HORARIO
+            if(getHorarios(id)) {
+                return true;
+            }else{
+                return false;
+            }
+        } catch (Exception e) {
+            return false;
+        }
     }
 
     //Metodo que obtiene el horario de marcación del trabajador
     //para validar la hora de asistencia
-    private void getHorarios(int idUser) {
-        horariosArray = new HorarioViewModel();
-        //OBTENIENDO DIA ACTUAL
+    private boolean getHorarios(int idUser) {
+        final boolean[] result = {true};
+        String ahora = DateFormat.getDateTimeInstance().format(new Date());
         Calendar calendar = Calendar.getInstance();
         calendar.setFirstDayOfWeek(Calendar.MONDAY);
         calendar.setMinimalDaysInFirstWeek(4);
         calendar.setTime(new Date());
+        int anio = calendar.get(Calendar.YEAR);
+        int mes = calendar.get(Calendar.MONTH);
+        int dia = calendar.get(Calendar.DAY_OF_MONTH);
         int numeroSemana = calendar.get(Calendar.DAY_OF_WEEK);
-        numeroSemana = numeroSemana == 0 ? 7 : numeroSemana-1;
+        numeroSemana = numeroSemana == 0 ? 7 : numeroSemana - 1;
         queue = Volley.newRequestQueue(getActivity());
-        Map<String, Integer> params = new HashMap();
-        params.put("diaActual", numeroSemana);
-        params.put("idTrabajador", idUser);
-        JSONObject jsonObject = new JSONObject(params);
-        JsonObjectRequest jsonObjectRequest = new JsonObjectRequest(Request.Method.GET, ApiURL + "Horario/ObtenerHorario", jsonObject, new Response.Listener<JSONObject>() {
+        StringRequest stringRequest = new StringRequest(Request.Method.GET, ApiURL + "Horario/ObtenerHorario?diaActual="+numeroSemana+"&idTrabajador="+idUser, new Response.Listener<String>() {
             @Override
-            public void onResponse(JSONObject response) {
-                //DESERIALIZANDO RESPUESTA
+            public void onResponse(String response) {
                 Gson gson = new Gson();
-                horariosArray = gson.fromJson(response.toString(), HorarioViewModel.class);
+                try {
+                    horariosArray = gson.fromJson(response, HorarioViewModel[].class);
+                    if (horariosArray.length == 0) {
+                        HoraInicio = "";
+                        HoraFin = "";
+                        Sigla = "";
+                        Tipo = "";
+                    }
+                    for (int x = 0; x < horariosArray.length; x++) {
+                        String hoy = anio + "-" + mes + "-" + dia + " " + horariosArray[x].getHodC_INICIO();
+                        HoraInicio = new SimpleDateFormat("YYYY-MM-DD HH:mm:ss").parse(anio + "-" + mes + "-" + dia +  " " + horariosArray[x].getHodC_INICIO(), new ParsePosition(0)).toString();
+                        HoraFin = new SimpleDateFormat("YYYY-MM-DD HH:mm:ss").parse(anio + "-" + mes + "-" + dia + " " + horariosArray[x].getHodC_FIN(), new ParsePosition(0)).toString();
+                    }
+                        } catch (Exception e) {
+                    new SweetAlertDialog(getActivity(), SweetAlertDialog.ERROR_TYPE).setTitleText(e.getMessage()).show();
+                    result[0] = false;
+                }
             }
         }, new Response.ErrorListener() {
             @Override
@@ -153,7 +182,8 @@ public class HomeFragment extends Fragment {
                 Toast.makeText(getContext(), error.getMessage(), Toast.LENGTH_SHORT).show();
             }
         });
-        queue.add(jsonObjectRequest);
+        queue.add(stringRequest);
+        return result[0];
     }
 
 }
